requirements.txt
# Core
fastapi
uvicorn
pydantic
python-dotenv
PyYAML

# Async and job queue
httpx
celery
redis

# Security
passlib
python-jose

# LLM and NLP
langchain
ollama

# Utilities
requests
loguru

# Frontend (for dev)
Jinja2

# Optional: for plugin loading
importlib-metadata; python_version<'3.8'